requirements.txt
Django==2.2.2
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-db-signals==0.1.1
django-js-asset==1.2.2
Pillow==6.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
